refactor: 카드 정보 state를 초기화하는 함수가 카드 등록 완료 페이지 이동 전에 호출되도록 수정
- 뒤로가기 시, 폼 데이터를 초기화 하기 위해 로직 수정
