+++ src/components/registerCompletePage.tsx/RegisterComplete.tsx
@@ -1,42 +1,25 @@
-import { useNavigate } from 'react-router-dom';
+import { useLocation, useNavigate } from 'react-router-dom';
 import Button from '../common/button/Button';
-import { INITIAL_CARD, ROUTER } from '../../global/constants';
+import { ROUTER } from '../../global/constants';
 import styles from './RegisterComplete.module.css';
-import useCardContext from '../../hooks/useCardContext';
 
 function RegisterComplete() {
-  const {
-    cardNumbers,
-    cardIssuer,
-    setCardNumbers,
-    setCardIssuer,
-    setExpiryDate,
-    setCardCVC,
-    setPassword,
-  } = useCardContext();
   const navigate = useNavigate();
+  const location = useLocation();
+  const { state } = location;
 
   function navigateMainPage(e: React.MouseEvent<HTMLButtonElement>) {
     e.preventDefault();
 
-    initForm();
     navigate(ROUTER.main);
-  }
-
-  function initForm() {
-    setCardNumbers(INITIAL_CARD.cardNumbers);
-    setCardIssuer(INITIAL_CARD.cardIssuer);
-    setExpiryDate(INITIAL_CARD.expiryDate);
-    setCardCVC(INITIAL_CARD.cardCVC);
-    setPassword(INITIAL_CARD.password);
   }
 
   return (
     <section className={styles.section}>
       <div className={styles.container}>
         <img src='../CompleteIcon.png' alt='등록 완료 아이콘' />
-        <p className='tx-xl'>{cardNumbers[0]}로 시작하는</p>
-        <p className='tx-xl'>{cardIssuer}가 등록되었어요.</p>
+        <p className='tx-xl'>{state.cardNumbers[0]}로 시작하는</p>
+        <p className='tx-xl'>{state.cardIssuer}가 등록되었어요.</p>
         <Button
           type='button'
           name='goToMainButton'
